Abandon the custom DieTree model
Use the canonical GtkTreeStore instead.  The same lazy child-expansion
is achieved with a placeholder for any dwarf_haschildren, which gets
filled in by a callback on the GtkTreeView's "row-expanded" signal.

In theory it's more wasteful for every node to carry all the Dwarf_Die-
generated metadata, instead of producing those on request, but I didn't
measure any real difference in memory usage.