Generate the known-dwarf.h macros
The previous known-dwarf.h was "stolen" from an elfutils build, but the
problem is that the constants within will depend on some minimum version
of dwarf.h.  A better solution is to steal elfutils' known-dwarf.awk
generating script instead, so the string tables will be generated from
the exact dwarf.h on the dwarvish build system.

This commit also adds dwstring.c to provide complete generated coverage
of all dwarf.h constant types.