Store just a boxed Dwarf_Die in the tree
Having the original Dwarf_Die is easier than carrying the base Dwarf and
types bool around everywhere to get a Dwarf_Die from an offset.  This
also goes back to using a custom data function for cell renders, rather
than storing so much text in the tree all the time.